"when freedom patches fight, nobody wins" -@jeremy
Use `JSON.generate` instead of `Object#to_json`

The reasoning behind this is that `Object#to_json` is part of the
[freedom]((https://github.com/rails/rails/blob/master/activesupport/lib/active_support/core_ext/object/to_json.rb#L7-L19)
[patching](https://github.com/rails/rails/blob/master/activesupport/lib/active_support/json/encoding.rb#L305-L312)
wars with `ActiveSupport`. This means that in Rails out of the box for
Ruby 1.9+ (since json is part of stdlib), the asset pipeline will use
the pure ruby `#to_json` method provided by `ActiveSupport`. This
results in ~40% performance decrease.