Cache options generated in MultiJson::Adapter
Both MultiJson::Adapter.dump and .load accept an options parameter
which had been processed on every call to before being passed through
to the underlying adapater. However, the options rarely change and
so much of this work was redundant. This changes the class to
cache the processed options, this saves a small amount of work per
call to .dump or .load.