Wrap dynamic require in mutex
`require` is not thread safe in Jruby. For more detail see:

    https://github.com/jruby/jruby/wiki/Concurrency-in-jruby#thread_safety

In our Rails project we use jbuilder (https://github.com/rails/jbuilder)
to generate JSON responses to XHRs. On one particular page we issue about
20 or 30 requests in rapid succession. Occasionally this `require` fails and
we see:

    uninitialized constant MultiJson::Adapters
    [RAILS_ROOT]/gems/gems/multi_json-1.12.1/lib/multi_json.rb:159:in
    `load_adapter_from_string_name'

From the logs it appears to happens when the adapter load is triggered by
two threads simultaneously. The failure is avoided by wrapping the require
in a Mutex.